Added debug user for logging in when "Master" is missing.

+++ NTDLS.Katzebase.Shared/Helpers.cs
@@ -50,16 +50,19 @@ namespace NTDLS.Katzebase.Shared
         {
             Directory.CreateDirectory(destinationPath);
 
-            //Now Create all of the directories
-            foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
+            if (Path.Exists(sourcePath))
             {
-                Directory.CreateDirectory(dirPath.Replace(sourcePath, destinationPath));
-            }
+                //Now Create all of the directories
+                foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
+                {
+                    Directory.CreateDirectory(dirPath.Replace(sourcePath, destinationPath));
+                }
 
-            //Copy all the files & Replaces any files with the same name
-            foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
-            {
-                File.Copy(newPath, newPath.Replace(sourcePath, destinationPath), true);
+                //Copy all the files & Replaces any files with the same name
+                foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
+                {
+                    File.Copy(newPath, newPath.Replace(sourcePath, destinationPath), true);
+                }
             }
         }
